requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.10
numpy>=1.20.0
requests>=2.25.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0
openai>=1.10.0

pyaudio
librosa
noisereduce
openai
SpeechRecognition
pyAudioAnalysis
numpy
openai-whisper
soundfile
scipy
numba
